Optimize ADPCM (sosCODEC) decompression
This commit optimizes ADPCM decompression in game by:

* Removing any unsupported modes that the game do not use.
* Using a table of possible values instead of computing them on each
  iteration.
* Generating a custom function for 16 bit stereo and mono by using
  a template function.

This new table is built by employing a dynamic programming techinque.
The possible values are bound and the product of the dimensions are
small, hence this techinque is quite effective in optimizing things
by precomputing the values.

Both mono 8-bits and stereo 8-bits function are not generated as no
sample ingame uses them.

Signed-off-by: Giuliano Belinassi <gbelinassi@suse.de>