Added a define for activating support for 'new' video code.
Vanilla Conquer had internal changes to how it handled frame
rendering when SDL support was added. This code is only active for
SDL builds, and not for builds with the legacy Windows video code
enabled. However, this code should be used for all new video code
implementations and not just SDL, so hiding it behind the SDL_BUILD
is misleading and also makes it harder to add support for other
platforms.

Changed the compile time checks for SDL_BUILD into NEW_VIDEO_BUILD
where those checks guard code that is not SDL-specific.